requirements.txt
langchain
openai
chromadb==0.3.29
langchainhub
python-dotenv
bs4
pysqlite3-binary
tiktoken
black
google-cloud-aiplatform>=1.35.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
unstructured
# pdf2image
# pdfminer
pypdf
psycopg
psycopg-binary